Refactor Models tab components into presentational and container components
