+++ static/js/components/Main.jsx
@@ -17,7 +17,7 @@ import ProjectSelector from '../containers/ProjectSelector';
 import ProjectTab from '../containers/ProjectTab';
 import DatasetsTab from './DatasetsTab';
 import FeaturesTab from '../containers/FeaturesTab';
-import ModelsTab from './Models';
+import ModelsTab from './ModelsTab';
 import PredictTab from './Predictions';
 import colorScheme from './colorscheme';
 import Progress from './Progress';
